fix to line numbers for errors on multilocale
Made an error ("illegal use of function") point to user code in
function resolution like many other errors in function resolution and
made it so that functions marked as compiler temporaries are not
thought of as user code (instead reporting at a call site).  Marked
coercion wrappers as compiler temporaries.  The same is not done for
default wrappers since user code (the default value assignments) is
inserted into the default wrapper.

This fixes the following GASNet regression:

  classes/bradc/overloadMethods/v2NoReturnValueButType


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13518 3a8e244f-b0f2-452b-bcba-4c88e055c3ca