Attempt to fix "gen_release" script for Mac
(Since gen_release uses the committed code, can't test
this very well until it's been committed).

gen_release wasn't working on the Mac because the default
shell for Makefiles on the Mac is "sh" and the version of
sh that it implements has an echo command that doesn't
like using the -n flag (which our copyright-tagging rules
rely on).  Added a platform-specific Makefile for the Mac
that uses bash as the shell instead.  If for any reason
this turns out to be unsatisfactory, we could also define
ECHO to refer to an echo binary, which would bypass the
version used by the shell.  For some reason that seemed
clumsier to me which is why I opted to go this route.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13527 3a8e244f-b0f2-452b-bcba-4c88e055c3ca