fix to inout/out args and class reads on multilocale
Fixed a bug in which updates to inout and out arguments would not be
reflected back at the call site if the argument was put on the heap
because it was used in either an on- or begin-statement.

Updated the default class read function to do the read on the locale
that the file exists on.  This fixes reading classes from files.

Removed two futures.  Updated the class read test so that the class is
constructed before it is read.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13528 3a8e244f-b0f2-452b-bcba-4c88e055c3ca