Removing dead command from gen_release
gen_release used to remove the (empty) hpcc06 directory from
the examples/directory.  This directory no longer exists
under subversion.

With this checkin, gen_release ought to work on the Mac,
with the exception that "latex" does not seem to be
installed on our Mac, so the latex document(s) that the
script builds will not be built.  This isn't a big deal,
though, it just means we don't bundle our official
releases on the Mac.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13531 3a8e244f-b0f2-452b-bcba-4c88e055c3ca