Making updateSVNLOG.pl update to -rHEAD
This change makes the updateSVNLOG.pl update your SVNLOG
file up to the latest checked-in version rather than the
copy that you have in your directory.  I kept the old
functionality around as well -- it can be changed by 
flipping a variable as indicated in the comments.

Let me know if anyone runs into problems with this new
version.  I've only tested it a little.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13533 3a8e244f-b0f2-452b-bcba-4c88e055c3ca