Fixing updateSVNLOG off-by-one duplication bug
I'm surprised I didn't catch this before, but it seems as though
the updateSVNLOG.pl script has duplicated entries when the SVNLOG
file is incrementally updated.  The problem was due to an off-by-one
error, and should be fixed with this checkin.

As always, let me know if you find problems with this script.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13535 3a8e244f-b0f2-452b-bcba-4c88e055c3ca