added checking of returning values or types
Added checks to ensure that a type function does not return a valuea
and a value function does not return a type.

Moved the code to check resolved moves for a move from a value to a
type and a move from a type to a value before a line that disables
checking resolved moves to return symbols as suggested by a comment.
Specialized the error message for returns.

Added support for using FnSymbol::canType to indicate that a function
can return a type or not.  This is used on the function that the
compiler builds to implement if-expressions.

Updated .goods and removed two futures.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13536 3a8e244f-b0f2-452b-bcba-4c88e055c3ca