Add a that defines a function strangely.
This test defines and uses function in a "for param" loop. The function is
cloned for each iteration, and only accessible from within its iteration of
the loop.

This is the correct behavior, but for me it brings up the question of if we
really want to allow defining a function in every arbitrary block.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13542 3a8e244f-b0f2-452b-bcba-4c88e055c3ca