type/value parsing improvements
Unified parsing of types and expressions such that any
statement-level-expression is accepted where a type is parsed.  The
notion of a statement-level-expression is internal and we should lift
this restriction if we plan to allow any expression to be parsed where
a type is needed.  The statement-level-expression is used to handle
the case of

  if expr then expr else expr;

since this could be parsed as both an if-statement or an
expression-statement where the expression in the expression-statement
is an if-expression.  Since the semantics are the same at the language
level, this is used only in the implementation to avoid ambiguities in
the YACC grammar.

Changed two .good files

  test/arrays/dinan/no_element_type_error.good
  test/arrays/bradc/noelemtype.good

as the error message "invalid type specification" for code that
specifies the type of a variable as an array type without an element
type (only allowed for formals) is more general, but still acceptable,
than it was before this change.

Made the "invalid type specification" halt compilation immediately as
opposed to continuing to find other errors.  Otherwise we seem to
start to give fairly obtuse error messages in the above tests.  Update
one .good file as a result of this change

  test/types/scalar/bradc/valAsType.good

Fixed four tests

  test/trivial/deitz/test_negate_uint64.chpl
  test/trivial/deitz/test_negate_uint32.chpl
  test/trivial/deitz/test_int32.chpl
  test/trivial/marybeth/casts1.chpl

which were getting away with breaking the precedence rules and writing
things like

  var ui : uint(64) = 2:uint(64)**45;

instead of

  var ui : uint(64) = (2:uint(64))**45;

Fixed one test

  test/types/tuple/deitz/test_htuple6.good

that declared a 1-tuple by parenthesizing a type.  With types working
more like expressions even in type contexts, the parentheses are
ignored.  This seems more consistent to me as the parentheses would be
ignored were the parenthesized type passed to a function that expected
a type.  Updated the tuple module to avoid the same kind of code.
Also overloaded * on an int (but not a param int) and a type to issue
the error that a "tuple size must be static".

Set a precedence for the "of" keyword.  This is a syntactic sugar that
we didn't prioritize before because it could only be used in a type
context.  If we allow types anywhere expressions are parsed, it needs
a precedence.  Even with allowing "*" anywhere types are used requires
this too.  There was previously no way to write "3*(list of int)".  I
set the precedence to be identical to ** because we want it to be
right-associative too.  I think the right thing to do is to remove
this keyword from the language.  It is a syntactic sugar that was
designed with sequences in mind but sequences are no longer part of
the language.

Removed two futures related to parsing homogeneous tuple types in the
context of another type, e.g., an array of homogeneous tuples.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13545 3a8e244f-b0f2-452b-bcba-4c88e055c3ca