Improve my distributed array class
In the my previous commit message, I stated that most array references in this
test should be local. Unfortunately, I didn't notice that although the array
references were local, it took at least two remote references to find the
location to locally reference.

I added one copy of the array data structure per locale, with the references
used to calculate the correct array location on that locale. This resulted in
a huge speedup, though it's still unbearably slow.

Also, changed an array copy used to swap arrays in between iterations into
a use of the swap operator.

At the tiny array size of 50, with GASNET_SPAWNFN=l, and 2 locales, the
previously committed version took 85 seconds. Now, it takes 2.4 seconds.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13547 3a8e244f-b0f2-452b-bcba-4c88e055c3ca