disable allocating local parameters on the heap
Added code to disable adding any local parameter to the heap in
normalization.  Also disabled adding any local variable with the
private pragma to the heap.  These checks were already done for
globals.

Added a special check later to disable adding any index of a parameter
loop to the heap.  These are not marked as parameters so this required
its own special check.

Removed two futures.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13548 3a8e244f-b0f2-452b-bcba-4c88e055c3ca