Random module improvements
Added a skipToNth() method to the RandomStream class that allows
the stream to be moved to the nth value without returning it
(as getNth() does).  This makes it simpler to get the stream to
a certain position before calling getNext() in a loop or
fillRandom().

Added a new test of the Random module that tests a number of
ways of calling getNth(), fillRandom(), and skipToNth() to
ensure that they return consistent values.

Updated the spec to document this new routine as well as
getNth() and getNext() which apparently never got documented.

Used these features to get stream-fragmented.chpl working and
in a simpler way than it would have needed to be under the
old interface.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13553 3a8e244f-b0f2-452b-bcba-4c88e055c3ca