Filing bug: simple read of sync var array element doesn't empty
This is similar to the test that showed that a simple read of
a sync var didn't empty it except that the sync var is not so
simple -- it's a sync element of an array rather than a plain
sync var.

Put in a test with halt to avoid making this test timeout.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13554 3a8e244f-b0f2-452b-bcba-4c88e055c3ca