Improved synchonization for fragmented stream's cooperative printing
Inspired by David's example earlier today (and spurred on
by Steve's disgust that I was spin-waiting on a readXX()),
optimized the synchronization used to take turns printing
out an array by using an array of sync vars whose fullness/
emptiness indicates whose turn it is rather than a single
sync var whose value indicates whose turn it is.

For a 100 element problem size on 4 locales (on my 2 core
processor) with --printArrays on, this reduced the runtime
from 65 seconds to 1 second.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13555 3a8e244f-b0f2-452b-bcba-4c88e055c3ca