removed support for some pragma syntax
Removed the ability to use pragmas in places in the code where we
haven't really needed to use them and where we don't really fully
support them.  For example, on expressions.

Removed tests of such pragma syntax where the test parsed the pragma
and then ignored it.  There are other such tests that handle the cases
we care about.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13557 3a8e244f-b0f2-452b-bcba-4c88e055c3ca