Finalizing draft of stream-fragmented.chpl
Removed debugging code and other cruft from stream-fragmented.chpl
and added its .numlocales file.   This finalizes this draft and
I'm going to move on to a more class-based implementation of
arrays next.

Am realizing that one downside to our current NUMLOCALES/.numlocales 
strategy is that it doesn't allow a test to be tested under multiple
locales for GASNet and a single locale for comm=none (which doesn't
make sense for all tests -- like those that refer to Locales(1) --
but it does make sense for others like this one).  Maybe we should
rely on the SKIPIF capability to lock comm=none out of certain tests?
Or, any other suggestions?

This test also makes me want to be able to specify that numlocales
(or other configs?) is a random number in some range (e.g., 1..4
or 1..MAXLOCALES where MAXLOCALES is specified on the start_test
command line) since it would be nice to stress test this test
against different numbers of locales over time (since the number
shouldn't matter) to make sure that things are robust; and yet to
avoid forcing every execution of start_test to run this test 
multiple times for a finite set of possibilities.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13559 3a8e244f-b0f2-452b-bcba-4c88e055c3ca