Making stream-fragmented a performance test
Making the stream-fragmented test with --no-local into
a performance test to start a baseline number before we
start doing optimizations.

Removed the .numLocales directory because this
prevents performance testing on a single locale as
we currently do.  Will reinstate this after
implementing the proposed test script change to have
.numlocales files allow execution on COMM=NONE runs
(and using .skipif files to prevent testing for
tests that don't make sense on COMM=NONE runs).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13561 3a8e244f-b0f2-452b-bcba-4c88e055c3ca