Version of STREAM using a manually written distribution class
This is the closest I've come to writing a distributed array
in Chapel -- I implemented global and local classes to
implement 1D blocked domains and arrays and wrote Stream
in terms of them.  In a complete version of Chapel, the
classes I wrote would be part of a distribution class, but
trying to fit into that framework and the modules was too
scary to do in one step.

My domain classes support:
- these() iterators
- an ind2loc() function
- a newArray() method
- standard dom, numIndices, low, and high queries (note that
  "dom" is normally called "domain", yet this causes syntax
  errors when done here.

My array classes support:
- this() methods
- these() iterators
- writeThis() methods
- numElements methods

There's more work to do here (as indicated in the TODOs in
the file): 

* I haven't used any on clauses yet, so am just
  pretending to use multiple locales (though the numLocales
  variable is used to determine how many locales I'm pretending
  to use / local classes I'm allocating).

* I hard-coded my index types into the code globally rather
  than parameterizing the classes by index type (global and
  local)

* I had to use initialize methods to get around inability to
  initialize array members inline.  I believe this is a known
  issue, but will check after this commit.

* Need to make the target set of Locales a member of the class
  rather than assuming that Locales/LocalesDomain are used

* Question about whether there's a cooler way to do my user
  constructor of LocBlock1DDom

* Need to clean up the Randomization code


I also think I found a bug or two which I've captured but
will check in separately after trying to boil them down to
a smaller example and verifying that they're not redundant.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13571 3a8e244f-b0f2-452b-bcba-4c88e055c3ca