Adding zippered reference semantics future
This future is similar to functions/deitz/iterators/test_var_iterator2.
It shows that when zippering two var iterators together, the reference
semantics are lost.  Had forgotten/didn't realize this, and by the
time I realized what the problem was, I'd already written the test, so
figured I'd check it in.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13574 3a8e244f-b0f2-452b-bcba-4c88e055c3ca