Future: missing 'new' keyword results in confusing error message
This is a very low-priority future -- I find that when I forget
the new keyword, the message that's generated is sometimes very
confusing (easy for me to work around since I quickly realize
that I forgot the new keyword; maybe harder for a user).

The .good file is just a placeholder -- I'm open to other error
messages.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13576 3a8e244f-b0f2-452b-bcba-4c88e055c3ca