Changes to nightly script names
This change makes two naming changes to thinkgs in the nightly
script -- (1) the unique ID representing the platform configuration
uses : separators rather than - to make the separation between
identifiers that use dashes (e.g., cray-mta) clearer;  (2) the
temporary directory in which the tests are run is now given a
non-random name (i.e., the process ID is no longer used), and
a week's worth of them are saved to allow inspection of the
files in question, the .comp.out.tmp/.exec.out.tmp files, etc.
After a week, the same directories are reused again (the net
effect being that we use more space when things are working
well, but don't have the problem cleaning up after ourselves
that we've traditionally had when things fail).

Users who run cron jobs should update their copy of this file
to see it work, and should check /tmp for *.deleteme directories
that can now be removed.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13577 3a8e244f-b0f2-452b-bcba-4c88e055c3ca