disabled heap allocation for coforall
Disabled allocating local variables accessed in coforall statements on
the heap.  Since coforall is implemented using begin blocks, we were
allocating such variables on the heap, but since we know that the
coforall will complete before we return from the function, this is not
necessary.

I had a bit of trouble (indicated by a comment) trying to omit the
calls to _downEndCount and _upEndCount which should also be
unnecessary.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13579 3a8e244f-b0f2-452b-bcba-4c88e055c3ca