Sync var reads at statement level
Make reads of sync vars at statement level result in a readFE (or readFF for
singles). Previously this only happened for simple sync variables, now it also
happens when they are for example class or array members.

Did this by passing the return from these calls to the _statementLevelSymbol
no-op function after functions are resolved, but before post resolution pruning.

Remove the .futures that want this behavior, update Brad's stream test to
remove temps used to get this behavior.

Add a test that tries to create a tuple of sync bools that dies with C compiler
warnings about making a pointer from an integer.

Add a working test that has an array of classes with sync fields.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13582 3a8e244f-b0f2-452b-bcba-4c88e055c3ca