Changing separators for nightly script again :(
Turns out that : is not portable for file naming -- OSX and
Windows don't like it, and it also breaks something in the
build of GASNet.  After a long study of the psychology and
readability of various choices, ended up using "." for the
separator between compiler, platform, threads, comm choices
and "--" for the cross-compilation case.

Feel free to send feedback on these choices.  Preferably
before the tests run tonight.

Meanwhile, I'll do a little more testing on platforms that 
were having problems than I did this weekend -- I hadn't
anticipated that ":" could be such a fragile character...




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13583 3a8e244f-b0f2-452b-bcba-4c88e055c3ca