limited applicability of array/domain rank change
Updated _validRankChangeArg to require that the integral arguments be
passable to the index type of that array.  This was not previously
done so that rank change could apply even if indexing could not making
rank change a viable possibility in such a case.  This would then give
a domain's error message since the array and domain rank change share
some code.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13586 3a8e244f-b0f2-452b-bcba-4c88e055c3ca