Bug: default array argument use causes copy
This is an interesting bug that I just ran into in my
associative domain/array rewrite: having a function
with a default array argument in which that default
argument is used results in the argument getting
copied rather than passed in by reference as arrays
normally are.

Steve said he would take a look at this.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13588 3a8e244f-b0f2-452b-bcba-4c88e055c3ca