Fixed runtime TAGS file
Fixed the runtime Makefile to look at the runtime sources
when generating its TAGS file.

Note that the TAGS file only reflects the most recent
build -- so only the thread/comm layer that you most
recently built in a given directory.  We can change this
if this becomes an issue for people (but given that
nobody noticed the file was broken, I'm guessing it
may not be... :)




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13589 3a8e244f-b0f2-452b-bcba-4c88e055c3ca