Extensions to generated code TAGS file
With this change, the TAGS file created for the
generated code includes the TAGS for the runtime
so that one can hop from generated code over to
the runtime and back again while editing.

This change also includes support for ctags-based
generation for the generated code TAGS file, 
though see my note on this feature below.

Implemented this by copying the runtime/TAGS file
over from the runtime directory and then appending
to it.  In order to make this copy work, I had to
make the runtime TAGS file use absolute path
names, so updated the Makefile.include files to
make this happen.

Added new Makefile variables to make this work as
well:

* CHPL_TAGS_FILE (the name of the tags file)
* CHPL_TAGS_APPEND_FLAG (the flag to use to append
  to a tags file)

Note that on Cray machines, multiple copies of
ctags are installed -- one is in the emacs
module while the other is in the standard root-level
/bin hierarchy.  Unfortunately, these take
different flags.  I've geared the current check-in
more on the /bin version of ctags than the emacs
version (because it seems newer), but those who
actually use vim and this feature may want to
compare notes on which version they use and want
to support by default (one can always customize
their environment variables to support either one
as desired).  I believe that the main change
required to use the emacs version of ctags is to
change to append flag from -a to -u.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13594 3a8e244f-b0f2-452b-bcba-4c88e055c3ca