Associative domain/array rewrite to avoid _ddata classes
The original associative domain/array code was written using
_ddata classes due to our poor support for using arrays within
other structures.  This rewrite swaps arithmetic domains/arrays
in for the _ddata classes in order to get debugging benefits
such as bounds checking, and to make the implementation a little
saner to reason about.

More rewrites to this module coming soon.  Consider this a
checkpoint.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13599 3a8e244f-b0f2-452b-bcba-4c88e055c3ca