Quick fix to GASNet associative domains/arrays
This is a workaround to the seg fault issue that
my associative domain/array check-in of the other
day caused.  In Steve's absence, David determined
that it was due to a reference to a global
variable within a field's initializer (and therefore
the default constructor's default argument list),
so I worked around it by moving the global variable
in question (the _ps array) into the classes for
the domain and array.  Wasted space, but it makes
those pesky seg faults go away.

In the meantime, David had some small tests of
the behavior written up and will check those in
as bug futures under the --no-local flag.

Seems like a possible fix might be to move some
of those more complex initializers (or ones that
refer to more global state) into their own
functions as part of the normalization process?
But that's said with only the vaguest grasp on
the issue.

Meanwhile, I'm continuing to debug my complete
rewrite of the associative domain/array class
and will check that in once I get a few more
bugs out.

Note: this check-in does not fix the opaque
example test regression.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13611 3a8e244f-b0f2-452b-bcba-4c88e055c3ca