Add tests demonstrating a --no-local bug
These tests are meant to demonstrate a problem with using global tuple and
array members as default values for function arguments. When globals are
moved to the heap, calls to a "_heapAccess" function are added to uses of
the global.

The clearest demonstration of the bug is:
defaultValueGlobalTuple.chpl

In the fieldDefault* and recordField* tests, it's caused by an argument's
default value in the class or record's default constructor. I wrote these
before I understood that the problem occurs with functions in general, not
just constructors.

To cause the bug to happen, the default value needs to be a global tuple or
array member.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13612 3a8e244f-b0f2-452b-bcba-4c88e055c3ca