Added compilerWarning capability
Added a capability to have compilerWarning
statements much like the current compilerError
statements that we currently support.
Support is identical except that a USR_WARN
is generated rather than a USR_FATAL.  This
is particularly useful for putting debugging
"printfs" into the standard modules while
you're working on them (as long as you want
to print out strings and types :).

Added a description of compilerWarning to the
spec, added an example for both compilerWarning
and compilerError, and added the statements
to the vim and emacs modes.

Added two tests.

One thing for Steve to code review is that
I don't remove the compilerWarning() call
from the AST but rather let it fall through
to codegen where I generate a comment for
it.  I tried to remove it right after the
USR_WARN(), but it seemed to kill the
statements that followed it as well.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13613 3a8e244f-b0f2-452b-bcba-4c88e055c3ca