Help with svn:ignore settings
In this check-in, I'm adding a new utility script
that can be run in a test directory to add a.out,
output, and *.tmp to its svn:ignore properties.
The idea is to simplify the steps required to add
a new test directory, or to retroactively add
svn ignores to an existing one.

Applied this script to a few directories where
I had some lingering *.tmp files.

Also, added a missing exit() call to the 
updateSVNLOG.pl script that I noticed when
copying its query of CHPL_HOME.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13618 3a8e244f-b0f2-452b-bcba-4c88e055c3ca