Fixed Aniruddha LocTree as record bug
Aniruddha has had a bug filed for some time with us in
which strange things were happening when he declared
LocTree to be a record.  I've postulated that this was
due to a subtle bug in SingleLocaleAssociative, and
that seems to be the case -- my checkin from this afternoon
fixed this bug (once I got its .good file correct and took
out my debugging code).  This is a simpler version of the
complete version of madness from which it was taken that
was enough to expose the bug previously.  It was simplified
primarily to reduce compile and execution time.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13621 3a8e244f-b0f2-452b-bcba-4c88e055c3ca