Changed the names of the functions in chplsys.c so they don't start with an underscore: _bytesPerLocale -> chpl_bytesPerLocale _coresPerLocale -> chpl_coresPerLocale _maxThreads -> chpl_maxThreads _maxThreadsLimit -> chpl_maxThreadsLimit
Also changed the names of the functions in error.c:
_chpl_warning -> chpl_warning
_chpl_error -> chpl_error
_chpl_internal_error -> chpl_internal_error
_chpl_msg -> chpl_msg

The potential for symbol name clashing can be reduced by having the Chapel
compiler mangle the names of any symbols starting with chpl in Chapel source
code.  Third party software linked in with Chapel programs is unlikely to have
symbol names starting with "chpl_".

Changed how the primitives corresponding to the chpl_maxThreads() and
chpl_maxThreadsLimit() functions are defined such that no special code is
needed in expr.cpp to generate C code to call these functions.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13624 3a8e244f-b0f2-452b-bcba-4c88e055c3ca