fixed handling of formal's default value
Made it so that the default value is no longer copied when passing it
to a formal argument.  Limited the semantics so that the default value
must be "passable" to the formal argument (rather than assignable).

Added a new internal helper function in the modules called
_createFieldDefault that handles fields such that the 'default value'
used by the default constructor works.  As an example, the function

  def foo(A: [1..3] int = 0) { ...

is no longer legal because the default value '0' cannot be passed to
the formal argument A.  On the other hand, the class

  class C { var A: [1..3] int = 0; ...

is still legal because '0' can be assigned to the field.  The
_createFieldDefault function handles this case when creating the
default constructor.

There are a few cases where copies are still done, including with
intent inout, and maybe with generic types and sync variables.  I plan
to look into these cases further to define the semantics.

Removed a future.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13630 3a8e244f-b0f2-452b-bcba-4c88e055c3ca