Fixing GASNet build when make OPTIMIZE=1
Added -O0 to the list of flags for the GASNet runtime file
as any other optimization level causes GASNet to throw
an assertion error if it was compiled with debug enabled
(which is how we compile it).

Longer-term, we probably want to build multiple copies of
GASNet and link against the appropriate one depending on
the setting of the OPTIMIZE variable in the Makefile.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13641 3a8e244f-b0f2-452b-bcba-4c88e055c3ca