Add some module/class interaction tests and futures
Add some tests that use a class from another module without a "use" of that
module in scope. This is done by returning an instance of the class from a
function that does use the module. The class fields and methods are not found,
with the exception of the writeThis method, called through writeln.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13642 3a8e244f-b0f2-452b-bcba-4c88e055c3ca