Fixed arrays of arrays in classes
With much-needed help and tips from Steve, fixed the
bugs related to arrays of arrays in classes which
were failing due to a failure to copy the "elt"
(element type field) in the runtime array type
structure -- a problem when using A.type where A is 
an array of arrays.

TODO: remove futures after making sure they pass
tonight.

TODO: think about a rewrite that would be even
simpler.

TODO: file an A.type case as described above.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13652 3a8e244f-b0f2-452b-bcba-4c88e055c3ca