Removing array of array within class futures
Removing all the futures that were failing due to 
arrays of arrays within classes that are now working.

Added an --instantiate-max flag to make
genericArrayInClass-otharrs compile, and added a
"performance" future for a duplicate copy of this
test which asks the question whether or not that
instantiate-max flag ought to be required  (this
is very low priority and just a placeholder -- it
seemed weird to me, but I didn't look into it
further).




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13654 3a8e244f-b0f2-452b-bcba-4c88e055c3ca