fixed bug with on-statements where serial is true
Added an argument to chpl_begin so that a remotely spawned task from
an on-statement is always spawned regardless of the serial state.  It
was incorrectly assumed that the serial state would always be false in
the handler but since the handler is run by the main thread, this is
not the case.  This showed up in a code like

  serial true {
    on Locales(1) {
      on Locales(0) {

where the second on-statement calls chpl_begin in the context where
serial is set to true.

Removed a future.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13667 3a8e244f-b0f2-452b-bcba-4c88e055c3ca