Update a test that had an incorrect assumption.
Update a test that assumed that formal arguments were at the same scope as
locals. Formals are further out of scope from locals and from symbols injected
via use statements. Fix the .good to expect this, and remove the .future.

Also add a test expecting a local to be nearer in scope than symbols injected
by a use statement.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13676 3a8e244f-b0f2-452b-bcba-4c88e055c3ca