fix constructor bug with arrays of syncs
Specialized the creation of default wrappers for default constructors
so that an argument to the default wrapper (one for which a default is
not used) is copied using the special _createFieldDefault function
just like the default initialization expression for a field is copied.
Also modified the constructor so that the typeExpr of a formal
argument is used to create a default temp if it is not created with
the defaultExpr (which happens if the defaultExpr is gNil or not
specified, or the argument is of 'out' intent).

Disabled insertion of reindexing code for array formal arguments on
wrappers.  This was not previously necessary because the typeExpr was
removed before a wrapper was created.

Removed code to remove the typeExpr of an argument after it is used to
determine the type of the argument.

Removed code in the default constructor to copy an argument into a
field since this is now done in the default wrapper for a default
constructor.  This eliminates the need for the special _set_field
function, shifting the burden to _createFieldDefault.  Removed the
_set_field function.

Fixed a bug in the default constructor in which the type of an
argument for a field without a type but with an initialization
expression was set to the initialization expression rather than the
TYPEOF the initialization expression.

This change makes it so that we no longer deadlock in the constructor
for a class that has a field that is an array of sync variables.

Updated a test to use "new," fixed the output of the atomic warning in
a .good, and removed two futures.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13683 3a8e244f-b0f2-452b-bcba-4c88e055c3ca