fix constructor bug with calling methods
Specialized the default wrapper for a default constructor to allocate
its own copy of 'this' and then pass it into the constructor via the
'meme' argument much as is done by subclass constructors.

Added an unfortunate line of code: When building the default wrapper
for a default constructor, I set the type of the meme argument in the
constructor to be the type of 'this' explicitly.  I couldn't think of
a better way to set this.

Fixed a bug in which type specifiers of nested classes (as in c.D if D
is a nested class in C and c is an instance of C) would not be
resolved correctly because the type specifier D was not marked as not
requiring parentheses.  Allowed type specifiers to ignore this check.
Perhaps a better way to fix this is to mark the type constructor for
non-generic classes as being parenthesesless, but I don't think this
is important.

Improved the default 'write' function for a nested class so that it
does not print out the value of the 'outer' field.  Without this
improvement, test/classes/deitz/nested/test_nested1 (for example)
results in infinite output because the outer class "has" an instance
of its inner class which in turn "has" an instance of its outer class.

This checkin fixes six futures.  In each of these futures, a method is
called by the constructor, because an initialization expression for a
field calls a method, but the method does not touch any fields.  I
think there will be problems with accessing fields in such methods.

Updated two goods and removed six futures.

This checkin also should fix the --no-inline regressions from last
night as well as (many or all of) the baseline regressions (had they
actually run).

Changed an internal error about not being able to resolve a type to a
user error.  Added classes/deitz/test_method_initialize_error.chpl as
a test of an error in which a field in a class with an elided type is
initialized by a method in that class that returns the same field.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13684 3a8e244f-b0f2-452b-bcba-4c88e055c3ca