fix constructor bug with calling methods that access fields
Also, this fixes the normalization of loop-expression-array-types and
conditional expressions (as well as other expressions that are parsed
using nested functions) that appear in expressions that define fields
in a class.  Such nested functions are made into methods.

Fixed ClassType::addDeclarations to work on BlockStmts and DefExprs
without using collect_top_asts.  This keeps us from adding
parser-created nested functions like the above from being incorrectly
duplicated.

Specialized the default wrapper for a default constructor to set the
fields of the created class in the default wrapper as the arguments
are processed.  This is disabled for fields in a super class
(including fields in a super class that are shadowed in the class
(which required extra care)).

Fixed a bug in handling of initialization expressions without types.
We now do an explicit "_copy" of such expressions to determine the
type of the argument.  This is important for cases that showed up in
arrayInClassRecord2 where the loop expression needs to be copied to
determine that the type is an array.

Updated three goods and removed four futures.  The arrayInClassRecord
series did not have complete goods so I filled them in.  The only
semantic issue remaining with these tests is that in one case they
capture an array into a tuple before passing it to a function.  This
creates a copy rather than capturing a reference (as expected by the
old incomplete .good).  There already exist futures dealing with this
issue.  My latest thought is that the construction of a tuple should
capture the fields like arguments capture actuals.  So the array
should be a reference.  This makes a tuple more like a function than a
record, but I think it is a better semantic definition.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13686 3a8e244f-b0f2-452b-bcba-4c88e055c3ca