fix tuple of sync bug
When instantiating dtAny (elided formal argument type) with a sync or
single type, we "desync" the type and rely on the coercion wrapper to
insert a readFE or readFF for the argument.  Disabled this for tuple
constructors so that we can have tuples of syncs.  This is consistent
with the desire to move towards a semantics where tuples capture their
actuals like a function's formal arguments.  Note that this is also
disabled for heap and reference constructors.  Note also that I could
not rely on using the tuple pragma because it is removed after the
tuple constructor is instantiated once (tuple constructors are
instantiated twice, first for the size of the tuple and then for the
arguments).  Instead this relies on a name match of
"_construct__tuple".  Perhaps this should use its own pragma.

Removed a future.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13688 3a8e244f-b0f2-452b-bcba-4c88e055c3ca